Standardizes the canvas course copy fix to follow CourseCopyPlugin
Adapt the course copy fix for Canvas to the CourseCopyPlugin
format without any functional changes.

It should be possible to eventually share more code with the rest of the
LMS's but that will require data migrations and more careful testing.

This change aligns the code location so it matches the expectations set
by the plugin.
