+++ lms/product/canvas/_plugin/course_copy.py
@@ -1,5 +1,49 @@
+from lms.services.exceptions import FileNotFoundInCourse
+
+
 class CanvasCourseCopyPlugin:
     """Handle course copy in Canvas."""
+
+    def __init__(self, api, file_service):
+        self._api = api
+        self._file_service = file_service
+
+    def is_file_in_course(self, course_id, file_id):
+        """Raise if the current user can't see file_id in course_id."""
+        for file in self._api.list_files(course_id):
+            # The Canvas API returns file IDs as ints but the file_id param
+            # that this method receives (from our proxy API) is a string.
+            # Convert ints to strings so that we can compare them.
+            if str(file["id"]) == file_id:
+                return
+
+        raise FileNotFoundInCourse("canvas_file_not_found_in_course", file_id)
+
+    def find_matching_file_in_course(self, course_id, file_ids):
+        """
+        Return the ID of a file in course_id that matches one of the files in file_ids.
+
+        Search for a file that the current Canvas user can see in course_id and
+        that matches one of the files in file_id's (same filename and size) and
+        return the matching file's ID.
+
+        Return None if no matching file is found.
+        """
+        for file_id in file_ids:
+            file = self._file_service.get(file_id, type_="canvas_file")
+
+            if not file:
+                continue
+
+            for file_dict in self._api.list_files(course_id):
+                if (
+                    file_dict["display_name"] == file.name
+                    and file_dict["size"] == file.size
+                    and str(file_dict["id"]) != file.lms_id
+                ):
+                    return str(file_dict["id"])
+
+        return None
 
     def find_matching_group_set_in_course(self, _course, _group_set_id):
         # We are not yet handling course copy for groups in Canvas.
@@ -9,5 +53,8 @@ class CanvasCourseCopyPlugin:
         return None
 
     @classmethod
-    def factory(cls, _context, _request):
-        return cls()
+    def factory(cls, _context, request):
+        return cls(
+            api=request.find_service(name="canvas_api_client"),
+            file_service=request.find_service(name="file"),
+        )
