+++ lms/services/canvas.py
@@ -6,9 +6,9 @@ class CanvasService:
 
     api = None
 
-    def __init__(self, canvas_api, file_service):
+    def __init__(self, canvas_api, course_copy_plugin):
         self.api = canvas_api
-        self._finder = CanvasFileFinder(canvas_api, file_service)
+        self._course_copy_plugin = course_copy_plugin
 
     def public_url_for_file(
         self, assignment, file_id, course_id, check_in_course=False
@@ -32,7 +32,7 @@ class CanvasService:
 
         try:
             if check_in_course:
-                self._finder.assert_file_in_course(course_id, effective_file_id)
+                self._course_copy_plugin.is_file_in_course(course_id, effective_file_id)
             return self.api.public_url(effective_file_id)
         except (FileNotFoundInCourse, CanvasAPIPermissionError):
             # The user can't see the file in the course. This could be because:
@@ -45,7 +45,7 @@ class CanvasService:
             #
             # We'll try to find another copy of the same file that the current
             # user *can* see in the current course and use that instead.
-            found_file_id = self._finder.find_matching_file_in_course(
+            found_file_id = self._course_copy_plugin.find_matching_file_in_course(
                 course_id,
                 # Use a set to avoid searching for the same ID twice if file_id
                 # and effective_file_id are the same.
@@ -64,53 +64,8 @@ class CanvasService:
             return url
 
 
-class CanvasFileFinder:
-    """A helper for finding file IDs in the Canvas API."""
-
-    def __init__(self, canvas_api, file_service):
-        self._api = canvas_api
-        self._file_service = file_service
-
-    def assert_file_in_course(self, course_id, file_id):
-        """Raise if the current user can't see file_id in course_id."""
-        for file in self._api.list_files(course_id):
-            # The Canvas API returns file IDs as ints but the file_id param
-            # that this method receives (from our proxy API) is a string.
-            # Convert ints to strings so that we can compare them.
-            if str(file["id"]) == file_id:
-                return
-
-        raise FileNotFoundInCourse("canvas_file_not_found_in_course", file_id)
-
-    def find_matching_file_in_course(self, course_id, file_ids):
-        """
-        Return the ID of a file in course_id that matches one of the files in file_ids.
-
-        Search for a file that the current Canvas user can see in course_id and
-        that matches one of the files in file_id's (same filename and size) and
-        return the matching file's ID.
-
-        Return None if no matching file is found.
-        """
-        for file_id in file_ids:
-            file = self._file_service.get(file_id, type_="canvas_file")
-
-            if not file:
-                continue
-
-            for file_dict in self._api.list_files(course_id):
-                if (
-                    file_dict["display_name"] == file.name
-                    and file_dict["size"] == file.size
-                    and str(file_dict["id"]) != file.lms_id
-                ):
-                    return str(file_dict["id"])
-
-        return None
-
-
 def factory(_context, request):
     return CanvasService(
         canvas_api=request.find_service(name="canvas_api_client"),
-        file_service=request.find_service(name="file"),
+        course_copy_plugin=request.product.plugin.course_copy,
     )
